adds usePlugins hook to the discovery page
